REFACTOR: Clarify color image data file name
